#176 - Address pacing PR review: install config, fix nbytes/comment
- examples/CMakeLists.txt: add daqiri_bench_raw_tx_rx_pacing.yaml to
  DAQIRI_BENCH_CONFIGS so it is copied to build/examples/ and installed to
  /opt/daqiri/bin/. The documented run command referenced an uninstalled file.
- daqiri_dpdk_engine.h: correct the now_tx_ns() fallback comment (it said
  CLOCK_REALTIME but the implementation uses CLOCK_MONOTONIC, since the NIC PTP
  clock free-runs from driver load unless a PTP daemon disciplines it).
- types.h: widen BurstHeaderParams::nbytes to uint64_t and compute the
  set_all_packet_lengths total in uint64 so a large batch x frame size cannot
  overflow the per-burst byte total the pacing path reads.
- daqiri_dpdk_engine.cpp: clear hdr.nbytes when a TX metadata buffer is acquired
  (create_tx_burst_params / get_tx_metadata_buffer), so the pacing "nbytes == 0
  means not yet accumulated" sentinel stays correct on recycled buffers even if a
  caller fills packet lengths without calling set_header first.

+++ include/daqiri/types.h
@@ -115,7 +115,7 @@ struct BurstHeaderParams {
   uint16_t port_id;
   uint16_t q_id;
   int num_segs;
-  uint32_t nbytes;
+  uint64_t nbytes;
   uintptr_t first_pkt_addr;
   uint32_t max_pkt;
   uint32_t max_pkt_size;
